--- app/src/main/java/woowacourse/movie/model/ItemUI.kt
@@ -1,3 +1,0 @@
-package woowacourse.movie.model
-
-interface ItemUI
