+++ app/src/main/java/woowacourse/movie/ui/movielist/adapter/AdvertisementViewHolder.kt
@@ -4,7 +4,7 @@ import android.view.View
 import android.widget.ImageView
 import androidx.recyclerview.widget.RecyclerView
 import woowacourse.movie.R
-import woowacourse.movie.model.AdvertisementUI
+import woowacourse.movie.ui.movielist.model.AdvertisementUI
 
 class AdvertisementViewHolder(view: View) : RecyclerView.ViewHolder(view) {
     private val advertisement = view.findViewById<ImageView>(R.id.iv_advertisement)
