+++ app/src/main/java/woowacourse/movie/ui/movielist/model/ItemUI.kt
@@ -1,0 +1,3 @@
+package woowacourse.movie.ui.movielist.model
+
+interface ItemUI
